refactor(Player): getY() 리턴 타입 float으로 변경

+++ src/game/Player.cpp
@@ -211,7 +211,7 @@ int Player::getX() const {
     return x;
 }
 
-int Player::getY() const {
+float Player::getY() const {
     return y;
 }
